requirements.txt
# Core dependencies
paho-mqtt==2.1.0
cherrypy==18.8.0
requests==2.31.0
python-dotenv==1.0.0

# Database
influxdb-client==1.36.1

# Telegram Bot
python-telegram-bot==20.4

# Authentication
pyjwt==2.8.0

# Analytics dependencies
numpy==1.26.0
pandas==2.1.0
statsmodels==0.14.0

# Additional utilities
urllib3==2.0.4
httpx==0.25.0 